requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
SQLAlchemy==1.4.46
asyncpg==0.27.0
pydantic==1.10.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.24.0
beautifulsoup4==4.11.2
PyPDF2==3.0.1
transformers==4.29.2
torch==2.0.1
faiss-cpu==1.7.2
python-dotenv==1.0.0
selenium
webdriver-manager
sentence-transformers
pdfplumber
python-docx
requests